cs2 rama y resolución

+++ src/main/java/com/bootcamp/be_java_hisp_w25_g02/dto/response/FollowerCountDTO.java
@@ -1,7 +1,7 @@
 package com.bootcamp.be_java_hisp_w25_g02.dto.response;
 
 public record FollowerCountDTO
-        (Long user_id,
+        (Integer user_id,
          String user_name,
          Long followers_count
