+++ src/main/java/com/bootcamp/be_java_hisp_w25_g02/entity/User.java
@@ -13,7 +13,7 @@ import java.util.List;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class User {
-    Long user_id;
+    Integer user_id;
     String user_name;
     Boolean seller;
     List<Long> following;
